+++ pipelex/cogt/imgg/imgg_job_factory.py
@@ -4,7 +4,7 @@ from pipelex.cogt.imgg.imgg_job import ImggJob
 from pipelex.cogt.imgg.imgg_job_components import ImggJobConfig, ImggJobParams, ImggJobReport
 from pipelex.cogt.imgg.imgg_prompt import ImggPrompt
 from pipelex.config import get_config
-from pipelex.mission.job_metadata import JobCategory, JobMetadata
+from pipelex.pipeline.job_metadata import JobCategory, JobMetadata
 
 
 class ImggJobFactory:
