+++ pipelex/cogt/ocr/ocr_job_factory.py
@@ -4,7 +4,7 @@ from pipelex.cogt.ocr.ocr_input import OcrInput
 from pipelex.cogt.ocr.ocr_job import OcrJob
 from pipelex.cogt.ocr.ocr_job_components import OcrJobConfig, OcrJobParams, OcrJobReport
 from pipelex.config import get_config
-from pipelex.mission.job_metadata import JobCategory, JobMetadata
+from pipelex.pipeline.job_metadata import JobCategory, JobMetadata
 
 
 class OcrJobFactory:
