+++ pipelex/pipe_controllers/pipe_batch.py
@@ -14,8 +14,8 @@ from pipelex.core.stuff_factory import StuffFactory
 from pipelex.core.working_memory import WorkingMemory
 from pipelex.exceptions import PipeExecutionError
 from pipelex.hub import get_pipe_router
-from pipelex.mission.job_metadata import JobMetadata
 from pipelex.pipe_controllers.pipe_controller import PipeController
+from pipelex.pipeline.job_metadata import JobMetadata
 
 
 class PipeBatch(PipeController):
