+++ pipelex/pipe_works/pipe_job.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel, Field
 from pipelex.core.pipe_abstract import PipeAbstract
 from pipelex.core.pipe_run_params import PipeRunParams
 from pipelex.core.working_memory import WorkingMemory
-from pipelex.mission.job_metadata import JobMetadata
+from pipelex.pipeline.job_metadata import JobMetadata
 
 
 class PipeJob(BaseModel):
