+++ pipelex/pipe_works/pipe_job_factory.py
@@ -4,8 +4,8 @@ from pipelex.core.pipe_abstract import PipeAbstract
 from pipelex.core.pipe_run_params import PipeRunParams
 from pipelex.core.pipe_run_params_factory import PipeRunParamsFactory
 from pipelex.core.working_memory import WorkingMemory
-from pipelex.mission.job_metadata import JobMetadata
 from pipelex.pipe_works.pipe_job import PipeJob
+from pipelex.pipeline.job_metadata import JobMetadata
 
 
 class PipeJobFactory:
