+++ pipelex/pipe_works/pipe_router_protocol.py
@@ -3,8 +3,8 @@ from typing import Optional, Protocol
 from pipelex.core.pipe_output import PipeOutputType
 from pipelex.core.pipe_run_params import PipeRunParams
 from pipelex.core.working_memory import WorkingMemory
-from pipelex.mission.job_metadata import JobMetadata
 from pipelex.pipe_works.pipe_job import PipeJob
+from pipelex.pipeline.job_metadata import JobMetadata
 
 
 class PipeRouterProtocol(Protocol):
