+++ pipelex/pipeline/activity/activity_models.py
@@ -2,7 +2,7 @@ from typing import Any, Callable
 
 from pydantic import BaseModel
 
-from pipelex.mission.job_metadata import JobMetadata
+from pipelex.pipeline.job_metadata import JobMetadata
 
 
 class ActivityReport(BaseModel):
