+++ pipelex/pipeline/pipeline_manager.py
@@ -4,9 +4,9 @@ from pydantic import Field, RootModel
 from typing_extensions import override
 
 from pipelex.exceptions import PipelineManagerNotFoundError
-from pipelex.mission.pipeline_factory import PipelineFactory
-from pipelex.mission.pipeline_manager_abstract import PipelineManagerAbstract
-from pipelex.mission.pipeline import Pipeline
+from pipelex.pipeline.pipeline import Pipeline
+from pipelex.pipeline.pipeline_factory import PipelineFactory
+from pipelex.pipeline.pipeline_manager_abstract import PipelineManagerAbstract
 
 PipelineManagerRoot = Dict[str, Pipeline]
