+++ tests/pipelex/pipelex_asynch/test_pipe_running_variants.py
@@ -12,8 +12,8 @@ from pipelex.core.stuff_factory import StuffBlueprint
 from pipelex.core.working_memory import WorkingMemory
 from pipelex.core.working_memory_factory import WorkingMemoryFactory
 from pipelex.hub import get_mission_tracker, get_pipe_router, get_report_delegate
-from pipelex.mission.activity.activity_handler import ActivityHandlerForResultFiles
-from pipelex.mission.job_metadata import JobMetadata
+from pipelex.pipeline.activity.activity_handler import ActivityHandlerForResultFiles
+from pipelex.pipeline.job_metadata import JobMetadata
 from tests.pipelex.test_data import PipeTestCases
